it works now!
yay
