+++ itemmicroservice/src/main/java/com/service/ItemService.java
@@ -2,7 +2,9 @@ package com.service;
 
 import com.dao.ItemRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
+@Service
 public class ItemService {
 
     @Autowired
