+++ itemmicroservice/src/main/resources/application.properties
@@ -5,10 +5,13 @@ eureka.client.service-url.defaultZone=http://localhost:9091/eureka/
 #spring.datasource.url=jdbc:mysql://128.199.240.247/emma
 #spring.datasource.username=emma
 #spring.datasource.password=emmaDbPass
-spring.datasource.url = jdbc:mysql://localhost:3306/stock?zeroDateTimeBehavior=CONVERT_TO_NULL&serverTimezone=UTC
-spring.datasource.username = root
-spring.datasource.password = 23083015
-spring.jpa.hibernate.ddl-auto=update
+#spring.datasource.url = jdbc:mysql://localhost:3306/stock?zeroDateTimeBehavior=CONVERT_TO_NULL&serverTimezone=UTC
+#spring.datasource.username = root
+#spring.datasource.password = 23083015
+#spring.jpa.hibernate.ddl-auto=update
+
+spring.data.mongodb.uri=mongodb://167.71.211.249:27017/emma
+
 
 spring.application.name=items
